+++ src/CMakeLists.txt
@@ -114,6 +114,7 @@ add_library(webserv_lib STATIC
         lib/cgi/response.cpp
         lib/cgi/response.hpp
         lib/utils/types/either.hpp
+        lib/core/action/serveHttpAction.cpp
 )
 # lib/ を include path に追加
 # PUBLIC にすると、依存するターゲットにも反映される
